+++ src/css/getStyles.js
@@ -1,0 +1,6 @@
+define(function() {
+	function getStyles( elem ) {
+		return elem.ownerDocument.defaultView.getComputedStyle( elem, null );
+	}
+	return getStyles;
+});
