+++ src/css/var/rmargin.js
@@ -1,0 +1,3 @@
+define(function() {
+	return (/^margin/);
+});
