+++ src/css/var/rnumnonpx.js
@@ -1,0 +1,3 @@
+define(["../../var/pnum"], function(pnum) {
+	return new RegExp( "^(" + pnum + ")(?!px)[a-z%]+$", "i" );
+});
