+++ src/offset.js
@@ -1,9 +1,12 @@
 define([
 	"./core",
 	"./var/strundefined",
+	"./css/support",
+	"./css/var/rnumnonpx",
+	"./css/curCSS",
 	"./css",
 	"./selector" // contains
-], function( jQuery, strundefined ) {
+], function( jQuery, strundefined, support, rnumnonpx, curCSS ) {
 
 var docElem = window.document.documentElement;
 
@@ -178,5 +181,33 @@ jQuery.each( {scrollLeft: "pageXOffset", scrollTop: "pageYOffset"}, function( me
 	};
 });
 
+// Webkit bug: https://bugs.webkit.org/show_bug.cgi?id=29084
+// getComputedStyle returns percent when specified for top/left/bottom/right
+jQuery.each( [ "top", "left" ], function( i, prop ) {
+	jQuery.cssHooks[ prop ] = {
+		get: function( elem, computed ) {
+			if ( support.pixelPosition() ) {
+				// Hook not needed, remove it.
+				// Since there are no other hooks for prop, remove the whole object.
+				delete jQuery.cssHooks[ prop ];
+				return;
+			}
+
+			jQuery.cssHooks[ prop ].get = function ( i, prop ) {
+				if ( computed ) {
+					computed = curCSS( elem, prop );
+					// if curCSS returns percentage, fallback to offset
+					return rnumnonpx.test( computed ) ?
+						jQuery( elem ).position()[ prop ] + "px" :
+						computed;
+				}
+			};
+
+			return jQuery.cssHooks[ prop ].get( i, prop );
+		}
+	};
+});
+
 return jQuery;
+
 });
